kio_sftp: Use the right type for timeout_sec and timeout_usec.
libssh expects the values passed to the SSH_OPTIONS_TIMEOUT and
SSH_OPTIONS_TIMEOUT_USEC to be longs, not plain ints.

On 64-bit platforms with sizeof(long) > sizeof(int), this mismatch can
be problematic and potentially result in invalid memory access that
causes the calls to ssh_options_set() to fail.

Forward-ported from kde-runtime/3dc39e92d34b3612e90f7a0b34d5d845a7af0b72.

CCBUG:      331674
REVIEW:     120900

+++ sftp/kio_sftp.cpp
@@ -527,7 +527,7 @@ bool sftpProtocol::sftpOpenConnection (const AuthInfo& info)
     return false;
   }
 
-  int timeout_sec = 30, timeout_usec = 0;
+  long timeout_sec = 30, timeout_usec = 0;
 
   qCDebug(KIO_SMTP_LOG) << "Creating the SSH session and setting options";
